salary_calculation/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cors==1.0.1
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
filelock==3.8.2
future==0.18.2
gevent==22.10.2
greenlet==2.0.1
idna==3.4
kombu==5.2.4
prompt-toolkit==3.0.38
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.6
redis==4.5.1
requests==2.28.1
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.3
tldextract==3.4.0
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.6
zope.event==4.5.0
zope.interface==5.5.2
